fix: move poll module to etc
