+++ src/index.ts
@@ -7,7 +7,6 @@ import { AutoroleModule } from './modules/autorole';
 import { EtcModule } from './modules/etc';
 import { HelpChanModule } from './modules/helpchan';
 import { PlaygroundModule } from './modules/playground';
-import { PollModule } from './modules/poll';
 import { ReminderModule } from './modules/reminders';
 import { RepModule } from './modules/rep';
 import { TwoslashModule } from './modules/twoslash';
@@ -29,7 +28,6 @@ for (const mod of [
 	EtcModule,
 	HelpChanModule,
 	PlaygroundModule,
-	PollModule,
 	ReminderModule,
 	RepModule,
 	TwoslashModule,
